backend/pyproject.toml
[project]
name = "echo-agent-backend"
version = "0.1.0"
description = "EchoAgent FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "PyYAML>=6.0.2",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.21.0",
    "python-multipart>=0.0.9",
    "openai>=1.51.0",
    "pypdf>=4.0.0",
    "httpx>=0.28.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
